@@ -1,29 +1,26 @@
-'use strict';
-var assert = require('assert');
-var gutil = require('gulp-util');
-var markdown = require('./');
+import test from 'ava';
+import Vinyl from 'vinyl';
+import m from '.';
 
-it('should compile Markdown to HTML', function (cb) {
-	var stream = markdown();
+test.cb('compiles Markdown to HTML', t => {
+	const stream = m();
 
-	stream.once('data', function (file) {
-		assert.equal(file.relative, 'fixture.html');
-		assert.equal(file.contents.toString(), '<p><em>foo</em></p>\n');
+	stream.once('data', file => {
+		t.is(file.relative, 'fixture.html');
+		t.is(file.contents.toString(), '<p><em>foo</em></p>\n');
 	});
 
-	stream.on('end', cb);
+	stream.on('end', t.end);
 
-	stream.write(new gutil.File({
+	stream.end(new Vinyl({
 		path: 'fixture.md',
-		contents: new Buffer('*foo*')
+		contents: Buffer.from('*foo*')
 	}));
-
-	stream.end();
 });
 
-it('should expose the marked object', function(){
-	assert.ok(markdown.marked);
-	assert.ok(markdown.marked.Renderer);
-	assert.ok(markdown.marked.lexer);
-	assert.ok(markdown.marked.parser);
-})
+test('exposes the marked object', t => {
+	t.truthy(m.marked);
+	t.truthy(m.marked.Renderer);
+	t.truthy(m.marked.lexer);
+	t.truthy(m.marked.parser);
+});
